+++ packages/artifacts/src/generate-artifacts.ts
@@ -5,8 +5,8 @@ import {
     nodeSerializer,
     serializeLanguages
 } from "@lionweb/core"
-import { libraryExtractionFacade, libraryModel } from "@lionweb/test/dist/instances/library.js"
-import { multiExtractionFacade, multiModel } from "@lionweb/test/dist/instances/multi.js"
+import { libraryModel, libraryReader } from "@lionweb/test/dist/instances/library.js"
+import { multiModel, multiReader } from "@lionweb/test/dist/instances/multi.js"
 import { libraryLanguage } from "@lionweb/test/dist/languages/library.js"
 import { multiLanguage } from "@lionweb/test/dist/languages/multi.js"
 import { shapesLanguage } from "@lionweb/test/dist/languages/shapes.js"
@@ -49,12 +49,12 @@ writeFileSync(diagramPath("library-gen.puml"), generatePlantUmlForLanguage(libra
 writeFileSync(diagramPath("library-gen.md"), generateMermaidForLanguage(libraryLanguage))
 console.log(`generated diagrams for Library M2`)
 
-writeJsonAsFile(instancePath("library.json"), nodeSerializer(libraryExtractionFacade)(libraryModel))
+writeJsonAsFile(instancePath("library.json"), nodeSerializer(libraryReader)(libraryModel))
 console.log(`serialized library M1`)
 
 saveLanguageFiles(languageWithEnum, "with-enum")
 
 saveLanguageFiles(multiLanguage, "multi")
 
-writeJsonAsFile(instancePath("multi.json"), nodeSerializer(multiExtractionFacade)(multiModel))
+writeJsonAsFile(instancePath("multi.json"), nodeSerializer(multiReader)(multiModel))
 console.log(`serialized multi-language M1`)