+++ packages/test/src/core/multi.test.ts
@@ -1,6 +1,6 @@
 import { deserializeSerializationChunk, nodeSerializer } from "@lionweb/core"
 import { libraryInstantiationFacade } from "../instances/library.js"
-import { multiExtractionFacade, multiModel } from "../instances/multi.js"
+import { multiModel, multiReader } from "../instances/multi.js"
 
 import { libraryLanguage } from "../languages/library.js"
 import { multiLanguage } from "../languages/multi.js"
@@ -8,7 +8,7 @@ import { deepEqual } from "../test-utils/assertions.js"
 
 describe("multi-language test model", () => {
     it("[de-]serialize multi-language model", () => {
-        const serializationChunk = nodeSerializer(multiExtractionFacade)(multiModel)
+        const serializationChunk = nodeSerializer(multiReader)(multiModel)
         const deserialization = deserializeSerializationChunk(
             serializationChunk,
             libraryInstantiationFacade,