+++ packages/test/src/utilities/measurer.test.ts
@@ -2,8 +2,8 @@ import { nodeSerializer } from "@lionweb/core"
 import { measure, readFileAsJson, writeJsonAsFile } from "@lionweb/utilities"
 import { join } from "path"
 
-import { libraryExtractionFacade, libraryModel } from "../instances/library.js"
-import { multiExtractionFacade, multiModel } from "../instances/multi.js"
+import { libraryModel, libraryReader } from "../instances/library.js"
+import { multiModel, multiReader } from "../instances/multi.js"
 import { libraryLanguage } from "../languages/library.js"
 import { multiLanguage } from "../languages/multi.js"
 import { deepEqual } from "../test-utils/assertions.js"
@@ -22,13 +22,13 @@ describe("metrics computation", () => {
     }
 
     it("works on library", () => {
-        const serializationChunk = nodeSerializer(libraryExtractionFacade)(libraryModel)
+        const serializationChunk = nodeSerializer(libraryReader)(libraryModel)
         compareWithFile(measure(serializationChunk, []), "library-no-languages.metrics.json")
         compareWithFile(measure(serializationChunk, [libraryLanguage]), "library-with-languages.metrics.json")
     })
 
     it("works on multi-language model", () => {
-        const serializationChunk = nodeSerializer(multiExtractionFacade)(multiModel)
+        const serializationChunk = nodeSerializer(multiReader)(multiModel)
         compareWithFile(measure(serializationChunk, []), "multi-no-languages.metrics.json")
         compareWithFile(measure(serializationChunk, [multiLanguage]), "multi-with-languages.metrics.json")
     })